# tests: optimize block_reward_reallocation and reorder tests in test_runner