block reward reallocation test has been split in two segments (same file) as it was the longest and would run for the entire duration of the test_runner on my local machine with high -j levels. 

Reward reallocation:
develop 3m2.562s -> this branch 2m27.697s

Entire test suite (j16 local):
develop 311 seconds ->254 seconds
Entire test suite (j32)  275 seconds -> 215 seconds