@UdjinM6 2021-12-03
How about another optimization https://github.com/UdjinM6/dash/commit/94fc41dac9b5160e1d280fa95b8f0a2524cc5fb3 or maybe even #4603?